Rename IndexingPlugin -> SearchPlugin
