--- packages/core/agent/src/plugins/indexing/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2023 DXOS.org
-//
-
-export * from './indexing-plugin';
