+++ packages/core/agent/src/plugins/search/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './search-plugin';
