+++ packages/core/agent/src/plugins/search/search-plugin.ts
@@ -11,7 +11,7 @@ import { Context } from '@dxos/context';
 import { type Query, type Subscription, type TypedObject } from '@dxos/echo-schema';
 import { invariant } from '@dxos/invariant';
 import { log } from '@dxos/log';
-import { SearchRequest, type SearchResponse } from '@dxos/protocols/proto/dxos/agent/indexing';
+import { SearchRequest, type SearchResponse } from '@dxos/protocols/proto/dxos/agent/search';
 import { type Runtime } from '@dxos/protocols/proto/dxos/config';
 import { type GossipMessage } from '@dxos/protocols/proto/dxos/mesh/teleport/gossip';
 import { ComplexMap } from '@dxos/util';
@@ -20,35 +20,23 @@ import { AbstractPlugin } from '../plugin';
 
 type Options = Required<Runtime.Agent.Plugins.Indexing>;
 
-type IndexDocument = {
-  id: string;
-  json: string;
-};
-
-type SpaceIndex = {
-  space: Space;
-  query: Query;
-  subscription: Subscription;
-};
-
 const DEFAULT_OPTIONS: Options = {
   enabled: false,
 };
 
-export class IndexingPlugin extends AbstractPlugin {
+export class Search extends AbstractPlugin {
   private readonly _ctx = new Context();
+  private _options?: Options = undefined;
+  private _index?: MiniSearch;
 
   private readonly _spaceIndexes = new ComplexMap<PublicKey, SpaceIndex>(PublicKey.hash);
-
-  private readonly _indexedObjects = new Map<string, IndexDocument>();
 
   private readonly _indexOptions = {
     fields: ['json'],
     idField: 'id',
   };
 
-  private _options?: Options = undefined;
-  private _index?: MiniSearch;
+  private readonly _indexedObjects = new Map<string, IndexDocument>();
 
   constructor(private readonly _indexPath?: string) {
     super();
@@ -58,13 +46,13 @@ export class IndexingPlugin extends AbstractPlugin {
         const { index, options } = JSON.parse(serializedIndex);
         this._index = MiniSearch.loadJS(index, options);
       } catch (error) {
-        log.warn('failed to load index from file', { error });
+        log.warn('Failed to load index from file:', { error });
       }
     }
   }
 
   async open(): Promise<void> {
-    log('opening indexing plugin...');
+    log('Opening indexing plugin...');
 
     invariant(this._pluginCtx);
     const config = this._pluginCtx.client.config.values.runtime?.agent?.plugins?.indexing;
@@ -77,7 +65,9 @@ export class IndexingPlugin extends AbstractPlugin {
 
     this._pluginCtx.client.spaces.isReady.subscribe(async () => {
       invariant(this._pluginCtx, 'Client is undefined.');
+
       const space = this._pluginCtx.client.spaces.default;
+
       const unsubscribe = space.listen('dxos.agent.indexing-plugin', async (message) => {
         log('received message', { message });
         await this._processMessage(message);
@@ -101,7 +91,7 @@ export class IndexingPlugin extends AbstractPlugin {
     if (this._indexPath) {
       invariant(this._index);
       const serializedIndex = JSON.stringify({ index: this._index.toJSON(), options: this._indexOptions });
-      log('saving index to file', { path: this._indexPath });
+      log('Saving index to file:', { path: this._indexPath });
       writeFileSync(this._indexPath, serializedIndex, { encoding: 'utf8' });
     }
   }
@@ -159,7 +149,7 @@ export class IndexingPlugin extends AbstractPlugin {
   }
 
   private async _processMessage(message: GossipMessage) {
-    if (message.payload['@type'] !== 'dxos.agent.indexing.SearchRequest') {
+    if (message.payload['@type'] !== 'dxos.agent.search.SearchRequest') {
       log.warn('Indexing plugin received unexpected message type.', { type: message.payload['@type'] });
       return;
     }
@@ -176,8 +166,7 @@ export class IndexingPlugin extends AbstractPlugin {
     return {
       results: results.map((result) => {
         return {
-          // TODO(burdon): Types should be independent of plugin (core echo).
-          '@type': 'dxos.agent.indexing.SearchResult',
+          '@type': 'dxos.agent.search.SearchResult',
           spaceKey: result.id.split(':')[0],
           objectId: result.id.split(':')[1],
           score: result.score,
@@ -189,3 +178,14 @@ export class IndexingPlugin extends AbstractPlugin {
     };
   }
 }
+
+type IndexDocument = {
+  id: string;
+  json: string;
+};
+
+type SpaceIndex = {
+  space: Space;
+  query: Query;
+  subscription: Subscription;
+};
