+++ packages/core/protocols/src/proto/dxos/agent/search.proto
